+++ wasm/sound/registers.ts
@@ -106,14 +106,14 @@ export function handledWriteToSoundRegister(offset: u16, value: u16): boolean {
   return false;
 }
 
-export function getChannelStartingVolume(channelNumber: i8): u8 {
+export function getChannelStartingVolume(channelNumber: i32): u8 {
   // Simply need to get the top 4 bits of register 2
   let startingVolume: u8 = getRegister2OfChannel(channelNumber);
   startingVolume = (startingVolume >> 4);
   return (startingVolume & 0x0F);
 }
 
-export function isChannelDacEnabled(channelNumber: i8): boolean {
+export function isChannelDacEnabled(channelNumber: i32): boolean {
   // DAC power is controlled by the upper 5 bits of NRx2 (top bit of NR30 for wave channel).
   // If these bits are not all clear, the DAC is on, otherwise it's off and outputs 0 volts.
   if(channelNumber !== 3) {
@@ -131,14 +131,14 @@ export function isChannelDacEnabled(channelNumber: i8): boolean {
   }
 }
 
-export function isChannelEnabledOnLeftOutput(channelNumber: i8): boolean {
+export function isChannelEnabledOnLeftOutput(channelNumber: i32): boolean {
   let registerNR51: u8 = eightBitLoadFromGBMemory(Sound.memoryLocationNR51);
   // Left channel in the higher bits
   let bitNumberOfChannel: u8 = (<u8>channelNumber - 1) + 4;
   return checkBitOnByte(bitNumberOfChannel, registerNR51);
 }
 
-export function isChannelEnabledOnRightOutput(channelNumber: i8): boolean {
+export function isChannelEnabledOnRightOutput(channelNumber: i32): boolean {
   let registerNR51: u8 = eightBitLoadFromGBMemory(Sound.memoryLocationNR51);
   // Left channel in the higher bits
   let bitNumberOfChannel: u8 = (<u8>channelNumber - 1);
@@ -147,7 +147,7 @@ export function isChannelEnabledOnRightOutput(channelNumber: i8): boolean {
 
 // Function to get 1st register of a channel
 // Contains Duty and Length
-export function getRegister1OfChannel(channelNumber: i8): u8 {
+export function getRegister1OfChannel(channelNumber: i32): u8 {
   if (channelNumber === 1) {
     return eightBitLoadFromGBMemory(Channel1.memoryLocationNRx1);
   } else if (channelNumber === 2) {
@@ -161,7 +161,7 @@ export function getRegister1OfChannel(channelNumber: i8): u8 {
 
 // Function to get 2nd register of a channel
 // Contains Envelope Information
-export function getRegister2OfChannel(channelNumber: i8): u8 {
+export function getRegister2OfChannel(channelNumber: i32): u8 {
   if (channelNumber === 1) {
     return eightBitLoadFromGBMemory(Channel1.memoryLocationNRx2);
   } else if (channelNumber === 2) {
@@ -175,7 +175,7 @@ export function getRegister2OfChannel(channelNumber: i8): u8 {
 
 // Function to get 3rd register of a channel
 // Contains Fequency LSB (lower 8 bits)
-export function getRegister3OfChannel(channelNumber: i8): u8 {
+export function getRegister3OfChannel(channelNumber: i32): u8 {
   if (channelNumber === 1) {
     return eightBitLoadFromGBMemory(Channel1.memoryLocationNRx3);
   } else if (channelNumber === 2) {
@@ -187,7 +187,7 @@ export function getRegister3OfChannel(channelNumber: i8): u8 {
   }
 }
 
-export function setRegister3OfChannel(channelNumber: i8, value: u8): void {
+export function setRegister3OfChannel(channelNumber: i32, value: u8): void {
   if (channelNumber === 1) {
     eightBitStoreIntoGBMemorySkipTraps(Channel1.memoryLocationNRx3, value);
   } else if (channelNumber === 2) {
@@ -201,7 +201,7 @@ export function setRegister3OfChannel(channelNumber: i8, value: u8): void {
 
 // Function to get 4th register of a channel
 // Contains Fequency MSB (higher 3 bits), and Length Information
-export function getRegister4OfChannel(channelNumber: i8): u8 {
+export function getRegister4OfChannel(channelNumber: i32): u8 {
   if (channelNumber === 1) {
     return eightBitLoadFromGBMemory(Channel1.memoryLocationNRx4);
   } else if (channelNumber === 2) {
@@ -213,7 +213,7 @@ export function getRegister4OfChannel(channelNumber: i8): u8 {
   }
 }
 
-export function setRegister4OfChannel(channelNumber: i8, value: u8): void {
+export function setRegister4OfChannel(channelNumber: i32, value: u8): void {
   if (channelNumber === 1) {
     eightBitStoreIntoGBMemorySkipTraps(Channel1.memoryLocationNRx4, value);
   } else if (channelNumber === 2) {
